Renamed componentMapping to importMapping
