+++ plugin/tsconfig.json
@@ -10,5 +10,5 @@
         "skipLibCheck": true,
     },
     "include": ["src"],
-    "exclude": ["**/node_modules/**/", "dist"]
+    "exclude": ["**/node_modules/**/", "dist", "**/*.test.ts"]
 }
